Fix Toast types
